requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.4.2
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.14.0
jsonschema>=4.17.0
typing-extensions>=4.5.0
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0